// API_RocketLeague_Project_FrontEnd/src/app/models/Ranks.ts
export enum Ranks {
  UNRANKED = "UNRANKED",
  BRONZE_I = "BRONZE_I",
  BRONZE_II = "BRONZE_II",
  BRONZE_III = "BRONZE_III",
  SILVER_I = "SILVER_I",
  SILVER_II = "SILVER_II",
  SILVER_III = "SILVER_III",
  GOLD_I = "GOLD_I",
  GOLD_II = "GOLD_II",
  GOLD_III = "GOLD_III",
  PLATINUM_I = "PLATINUM_I",
  PLATINUM_II = "PLATINUM_II",
  PLATINUM_III = "PLATINUM_III",
  DIAMOND_I = "DIAMOND_I",
  DIAMOND_II = "DIAMOND_II",
  DIAMOND_III = "DIAMOND_III",
  CHAMPION_I = "CHAMPION_I",
  CHAMPION_II = "CHAMPION_II",
  CHAMPION_III = "CHAMPION_III",
  GRAND_CHAMPION_I = "GRAND_CHAMPION_I",
  GRAND_CHAMPION_II = "GRAND_CHAMPION_II",
  GRAND_CHAMPION_III = "GRAND_CHAMPION_III",
  SUPERSONIC_LEGEND = "SUPERSONIC_LEGEND",
}